Splitted test.py into smaller modules
