+++ Tests/test_meshing.py
@@ -1,0 +1,49 @@
+# Unit tests meshing
+from FESTIM import FESTIM
+import fenics
+import pytest
+import sympy as sp
+
+
+def test_mesh_and_refine_meets_refinement_conditions():
+    '''
+    Test that function mesh_and_refine() gives the right
+    refinement conditions
+    '''
+    def create_subdomains(x1, x2):
+        class domain(FESTIM.SubDomain):
+            def inside(self, x, on_boundary):
+                return x[0] >= x1 and x[0] <= x2
+        domain = domain()
+        return domain
+
+    def mesh_parameters(ini, size, refs, pos):
+        param = {
+            "initial_number_of_cells": ini,
+            "size": size,
+            "refinements":  []
+        }
+        for i in range(len(refs)):
+            param["refinements"].append({"cells": refs[i], "x": pos[i]})
+        return param
+    refinements = [[[2, 3], [0.5, 0.25]], [[3, 11], [0.5, 0.25]]]
+    for i in range(len(refinements)):
+        param = mesh_parameters(2, 1, refinements[i][0], refinements[i][1])
+        mesh = FESTIM.mesh_and_refine(param)
+
+        mf1 = FESTIM.MeshFunction('size_t', mesh, 1)
+        mf2 = FESTIM.MeshFunction('size_t', mesh, 1)
+        subdomain1 = create_subdomains(0, refinements[i][1][1])
+        subdomain1.mark(mf1, 1)
+        subdomain2 = create_subdomains(0, refinements[i][1][0])
+        subdomain2.mark(mf2, 2)
+        nb_cell_1 = 0
+        nb_cell_2 = 0
+        for cell in FESTIM.cells(mesh):
+            cell_no = cell.index()
+            if mf1.array()[cell_no] == 1:
+                nb_cell_1 += 1
+            if mf2.array()[cell_no] == 2:
+                nb_cell_2 += 1
+        assert nb_cell_1 >= refinements[i][0][1]
+        assert nb_cell_2 >= refinements[i][0][0]
